LayoutTests/editing/inserting/insert-paragraph-between-text.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<div id="test" contenteditable="true">
<div contenteditable="false">1</div>tryToInsertLineBreakInThisLine</div>

<script>
Markup.description('Testcase for bug https://webkit.org/b/113007: Unable to insert a paragraph in between some text whose previous sibling is a non-editable block.\n'+
'The test has passed if three lines are displayed instead of two, with the last line consisting of the letter "e".');

var test = document.getElementById('test');
test.focus();

var selection = window.getSelection();
selection.collapse(test, test.childNodes.length);
selection.modify("move", "backward", "character");
document.execCommand("InsertParagraph");

Markup.dump(test);
</script>
</body>
</html>
