third_party/blink/web_tests/platform/generic/editing/inserting/insert-paragraph-between-text-expected.txt

Testcase for bug https://webkit.org/b/113007: Unable to insert a paragraph in between some text whose previous sibling is a non-editable block.
The test has passed if three lines are displayed instead of two, with the last line consisting of the letter "e".
| "
"
| <div>
|   contenteditable="false"
|   "1"
| "tryToInsertLineBreakInThisLin"
| <div>
|   "<#selection-caret>e"
